Refactor TeamFightEditPage and Dusk tests to reuse macros, simplify methods, and improve player search/validation handling.
